+++ cuid2_test.go
@@ -1,7 +1,6 @@
 package cuid2
 
 import (
-	"math/rand"
 	"testing"
 )
 
@@ -91,17 +90,83 @@ func TestSessionCounter(t *testing.T) {
 }
 
 func TestCreatingFingerprintWithEnvKeyString(t *testing.T) {
-	fingerprint := createFingerprint(rand.Float64, getEnvironmentKeyString())
-	if len(fingerprint) < MinIdLength {
-		t.Error("Could not generate fingerprint of adequate length")
-		t.Fatalf("Expected length to be at least %v, but got %v", MinIdLength, len(fingerprint))
+	mockRandomFunc := func() float64 { return 0.124816 }
+	mockEnvStr := "KEY1KEY2"
+	expectedFingerprint := "0auhbh4kdn02a87je880p4dyw34ov0y6xm5ics64ezqi9gymln0jhf5goqp74qt977dm2or7mdw8ohgyawyou6ojbhgma1fcb"
+
+	fingerprint := createFingerprint(mockRandomFunc, mockEnvStr)
+
+	if fingerprint != expectedFingerprint {
+		t.Errorf("Fingerprint did not match expected value.\nGot: %s, Expected: %s", fingerprint, expectedFingerprint)
 	}
 }
 
 func TestCreatingFingerprintWithoutEnvKeyString(t *testing.T) {
-	fingerprint := createFingerprint(rand.Float64, "")
-	if len(fingerprint) < MinIdLength {
-		t.Error("Could not generate fingerprint of adequate length")
-		t.Fatalf("Expected length to be at least %v, but got %v", MinIdLength, len(fingerprint))
+	mockRandomFunc := func() float64 { return 0.2481632 }
+	expectedFingerprint := "dzgrtvh0dz0agj9g7wy2ww2mc3g8ddogm550pug0qyvdl91ebq3sowoqzi7jyl35sznv4nne4dqbg4dl5c8yo9vfuf3znvojy"
+	fingerprint := createFingerprint(mockRandomFunc, "")
+
+	if fingerprint != expectedFingerprint {
+		t.Errorf("Fingerprint did not match expected value.\nGot: %s\nExpected: %s", fingerprint, expectedFingerprint)
+	}
+}
+
+func TestDeterminismOfGeneration(t *testing.T) {
+	testCases := []struct {
+		name string
+		length int
+		fingerprint string
+		counterStart int64
+		timeMs int64
+		randomFloat float64
+		expectedID string
+		expectedNextID string
+	}{
+		{
+			name: "Short ID with low random value",
+			length: 10,
+			fingerprint: "test-fingerprint",
+			counterStart: 0,
+			timeMs: 1751850060928,
+			randomFloat: 0.1,
+			expectedID: "c79ab4qwd8",
+			expectedNextID: "ctfxvev2em",
+		},
+		{
+			name:          "Long ID with high random value",
+			length:        32,
+			fingerprint:   "fruit-salad",
+			counterStart:  476782360,
+			timeMs:        1751850806018,
+			randomFloat:   0.8,
+			expectedID:    "uhqvhs8l0q5ub01c37pgwfqak5az4l2n",
+			expectedNextID: "u2c7rvhbd6evwn1vj69bye7tj8e7ou2m",
+		},
+	}
+
+	for _, testCase := range testCases {
+		t.Run(testCase.name, func(t *testing.T) {
+			g := &cuidGenerator {
+				length: testCase.length,
+				counter: NewSessionCounter(testCase.counterStart),
+				fingerprint: testCase.fingerprint,
+			}
+
+			mockRandomFunc := func() float64 {
+				return testCase.randomFloat
+			}
+
+			firstID := g.generate(testCase.timeMs, mockRandomFunc)
+
+			if firstID != testCase.expectedID {
+				t.Errorf("First ID generated did not match expected.\nGot: %s, Expected: %s", firstID, testCase.expectedID)
+			}
+
+			secondID := g.generate(testCase.timeMs, mockRandomFunc)
+
+			if secondID != testCase.expectedNextID {
+				t.Errorf("Second ID generated did not match expected.\nGot: %s, Expected: %s", secondID, testCase.expectedNextID)
+			}
+		})
 	}
 }